+++ tests/fixtures/parity/javascript/helpers.js
@@ -1,0 +1,3 @@
+function helper() {}
+function other() {}
+module.exports = { helper, other };